+++ packages/visual-grid-client/src/sdk/getAllResources.js
@@ -75,18 +75,17 @@ function makeGetAllResources({resourceCache, fetchResource, extractCssResources,
       }
 
       await Promise.all(
-        missingResourceUrls.map(url => {
-          const fetchOptions = getFetchOptions({url, referer, userAgent, proxySettings})
-          return fetchResource(url, fetchOptions)
-            .then(async resource =>
-              assignContentfulResources(resources, await processResource(resource)),
+        missingResourceUrls.map(async url => {
+          try {
+            const fetchOptions = getFetchOptions({url, referer, userAgent, proxySettings})
+            const resource = await fetchResource(url, fetchOptions)
+            return assignContentfulResources(resources, await processResource(resource))
+          } catch (err) {
+            logger.log(
+              `error fetching resource at ${url}, setting errorStatusCode to 504. err=${err}`,
             )
-            .catch(ex => {
-              logger.log(
-                `error fetching resource at ${url}, setting errorStatusCode to 504. err=${ex}`,
-              )
-              resources[url] = new RGridResource({url, errorStatusCode: 504})
-            })
+            resources[url] = new RGridResource({url, errorStatusCode: 504})
+          }
         }),
       )