+++ packages/visual-grid-client/src/sdk/getRenderMethods.js
@@ -3,6 +3,7 @@
 function getRenderMethods(renderWrapper) {
   const doGetRenderInfo = renderWrapper.getRenderInfo.bind(renderWrapper)
   const doRenderBatch = renderWrapper.renderBatch.bind(renderWrapper)
+  const doCheckResources = renderWrapper.checkResources.bind(renderWrapper)
   const doPutResource = renderWrapper.putResource.bind(renderWrapper)
   const doGetRenderStatus = renderWrapper.getRenderStatus.bind(renderWrapper)
   const setRenderingInfo = renderWrapper.setRenderingInfo.bind(renderWrapper)
@@ -10,6 +11,7 @@ function getRenderMethods(renderWrapper) {
   return {
     doGetRenderInfo,
     doRenderBatch,
+    doCheckResources,
     doPutResource,
     doGetRenderStatus,
     setRenderingInfo,